backend/pyproject.toml
[project]
name = "backendapp"
version = "0.1.0"
description = ""
authors = [
    {name = "amadoutidiallo",email = "kassadiallo@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (==0.115.0)",
    "uvicorn[standard] (==0.30.6)",
    "sqlalchemy (==2.0.35)",
    "psycopg2-binary (==2.9.10)",
    "python-dotenv (==1.0.1)",
    "pydantic (==2.9.2)",
    "pydantic-settings (==2.6.1)",
    "sqlmodel (>=0.0.27,<0.0.28)",
    "google-api-python-client (>=2.187.0,<3.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
